requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

# Embeddings + ML
sentence-transformers==3.0.1
numpy<2
faiss-cpu==1.7.4

# Parsing & preprocessing
beautifulsoup4==4.12.3
lxml==5.3.0
markdownify==0.13.1
pymupdf==1.24.10
pillow==10.4.0
pytesseract==0.3.13

# Markdown rendering for /view
markdown==3.7

# Utils
python-dotenv==1.0.1
sqlite-utils==3.36
httpx==0.27.2
tqdm==4.66.5
